refactor(indexer-framework): aborting commit watermark
## Description

In a recent change the commit watermark was made to perform an extra
iteration before shutting down, however this can cause stalls during
shutdown if the task fails to connect to its store.

This situation occurs in calls to `sui start` where the CLI is
responsible for running the database process as a child as well as the
RPC services. The database is wound down by the `Ctrl-C` signal, while
the RPC services are gracefully shutting down.

The change was originally made to cater to the checkpoint indexer (or
any use of the object store) where watermark updates are infrequent, so
it was useful to flush the watermarks before shutting down in response
to a cancellation.

However, at the time of that change, the cancellation token was
responsible for signaling "normal shutdown" (e.g. after having processed
all checkpoints), as well as interrupts (unexpected termination).

The hope is that by using separate methods to communicate these two
forms of exit, we can avoid adding this graceful shutdown support.

## Test plan

The `sui` CLI no longer hangs when made to wait for the RPC services to
gracefully shutdown.