[move][move-vm][bella-ciao] Replace some VM asserts with invariant violations (#24513)
## Description 

This aims to replace all of the remaining VM asserts with invariant
violations (except for in the tracer).

## Test plan 

Tests still pass

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] Indexing Framework: