[sui-pkg-alt] Sui pkg alt source validation (#24236)
## Description 

This PR re-enables the `sui-source-validation` crate, by fixing the code
and tests. It tries to keep the same parity with what existed before the
new pkg system, but it does not try to fix anything that was broken
before.

## Test plan 

Existing tests were fixed to run again correctly.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] Indexing Framework: