[docs] Improve documentation around tooling (#24527)
## Description 

This PR improves the documentation on Sui tooling in a few ways:
- it reorders the prerequisites as going through this installation on
Windows it felt more natural than the current order.
- it removes the Installation Details description that was wrong; when
using brew, suiup, or choco it will only install `sui` binary (or
whatever you ask it to install via suiup`, and all the others are not
really used in any way.
- it removes the description of suiup's internal usages of paths. The
reason for doing so is that it's better to point to `suiup`'s README
that will always be the ground truth, and this way we'll avoid having
information in two places that might get out of sync.


## Test plan 

Existing tests.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] Indexing Framework:

---------

Co-authored-by: Jessie Mongeon <133128541+jessiemongeon1@users.noreply.github.com>