Make withdraw_funds_from_object public (#24296)
## Description 

Make the Move API public for object balance withdraw.
Check in runtime to make sure it cannot be called if accumulator is not
enabeld.

## Test plan 

Added a test to check that when accumulator is disabled, one cannot
withdraw from objects.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] Indexing Framework: