[move][cli] Fix gas profiles for code that aborts (#24544)
## Description 

Traces for executions that aborted did not have a `CloseFrame` for every
`OpenFrame` emitted (which makes sense since the execution aborted).
However this meant that when constructing gas profiles, there wouldn't
be a matching close frame event for every open frame event in those
cases, and the resulting gas profile was incorrect.

This fixes this issue by seeing if there are any remaining open frames
at the end of the trace, and if so closing them out in the resulting
profile, so the resulting gas profile is valid.

## Test plan 

Tested locally + saving gas profiles for executions that abort to make
sure we capture the correct output for these and track it.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] Indexing Framework: