pick(24501): fix(indexer-framework): clap requires group name (#24686)
## Description

Clap `requires` clauses refer to other flags by their field name, not
how they appear as a flag.

## Test plan

The `clickhouse-sui-indexer` example failed to run because of this
issue, and now it runs.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [x] Indexing Framework: Fixes a bug related to command-line argument
parsing for indexers built with the framework.