display: Extract entrypoint
## Description

Add an entrypoint for Display v2 to parse a string as a `Chain` and use
it to extract some part of a Move Value. This also requires being able
to start the lexer in expression mode, (and preventing the lexer from
entering back into text mode).

This change also adopts a new pattern for Display v2 entrypoints.
Previously, they accepted the details of a root object and a store for
dynamic loads which they used to construct an `Interpreter`. Now they
accept a reference to the interpreter. This allows the `Extract`
entrypoint in particular to return values whose lifetimes are tied to
that interpreter (which would otherwise have been a temporary value in
the body of `Extract::extract`.

## Test plan

New unit tests:

```
$ cargo nextest run -p sui-display
```